Cast expanded FP8 scales to the compute dtype
The K-grouped expansion dispatched correctly but produced garbage output. All
four prompts returned only exclamation marks, and warm-up reported status 6,
while the kernel-dispatch failure count fell from thousands to zero. The layout
change worked and the numerics did not.

ResolveLinearWeightFormat declares the K-grouped FP8 scales as the compute
dtype. The checkpoint ships weight_scale_inv as BF16, and this run fell back to
FP16 compute because torch.cuda.is_bf16_supported is false on V100. TurboMind
was therefore told the scales were FP16 while the bytes were still BF16. Both
types are 16 bits wide, so no size check failed and the values were read as
reinterpreted bits.

Cast in FP8Format.expand before the repeat, so the copy carries converted
values, and thread the resolved compute dtype in from _build_resolver. AWQ,
GPTQ and CompressedTensor already cast their scales in normalize, and both
reference implementations do the same: 1Cat-vLLM calls .to(torch::kFloat16)
inside fp8_sm70_prepare, and marlin_v100 calls .to(layer.orig_dtype).

Patch-family ID: PHASE0B-FP8-PACK

+++ lmdeploy/turbomind/converter.py
@@ -47,7 +47,10 @@ def _build_resolver(model_format: str | None,
         formats.append(CompressedTensorFormat(block_in=group_size))
         dtype = torch.float16
     elif model_format == 'fp8':
-        formats.append(FP8Format())
+        # Pass the resolved compute dtype: when the K-grouped scale expansion
+        # is active, ResolveLinearWeightFormat declares the scales as the
+        # compute dtype, so the tensor must actually carry it.
+        formats.append(FP8Format(scale_dtype=dtype))
     elif model_format == 'mxfp4':
         formats.append(MXFP4Format())
     else:
